add sync to database module
